+++ src/app/app.module.ts
@@ -13,7 +13,6 @@ import { MatListModule } from '@angular/material/list';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { TestContentComponent } from './pages/test-content/test-content.component';
-import { ToolbarComponent } from './pages/toolbar/toolbar.component';
 import { SidenavLinkItemComponent } from './pages/sidenav-link-item/sidenav-link-item.component';
 
 
@@ -24,7 +23,6 @@ import { SidenavLinkItemComponent } from './pages/sidenav-link-item/sidenav-link
     MainPageComponent,
     PageNotFoundComponent,
     TestContentComponent,
-    ToolbarComponent,
     SidenavLinkItemComponent
   ],
   imports: [